+++ lib/docker_compose_env.rb
@@ -4,16 +4,21 @@ require 'yaml'
 # The main module for DockerComposeEnv, this provides a single "easy mode"
 # method for setting up the environment.
 module DockerComposeEnv
-  def self.setup!(env: ENV, file: 'docker-compose.yml')
+  def self.setup!(env: ENV, file: 'docker-compose.yml', process_env: {})
     config = YAML.safe_load(File.read(file))
 
     config.fetch('services', {}).keys.each do |service_name|
       (config['services'][service_name]['ports'] || []).each do |container_port|
+        compose_port_info = nil
+
         if /(\d+)\/udp$/ =~ container_port
           parsed_port = /(?<port>\d+)\/udp$/.match(container_port).named_captures['port']
 
           command = "docker-compose --file=#{file} port --protocol=udp #{service_name} #{parsed_port}"
-          compose_port_info = `#{command}`
+          Open3.popen3(process_env, command) do |stdin, stdout, stderr, wait_thr|
+            compose_port_info = stdout.read
+            wait_thr.value
+          end
 
           if (compose_port_info =~ /^0\.0\.0\.0\:\d+\s*$/) == 0
             if env["#{service_name.upcase}_HOST"].nil?
@@ -26,7 +31,10 @@ module DockerComposeEnv
           end
         else
           command = "docker-compose --file=#{file} port #{service_name} #{container_port}"
-          compose_port_info = `#{command}`
+          Open3.popen3(process_env, command) do |stdin, stdout, stderr, wait_thr|
+            compose_port_info = stdout.read
+            wait_thr.value
+          end
 
           if (compose_port_info =~ /^0\.0\.0\.0\:\d+\s*$/) == 0
             if env["#{service_name.upcase}_HOST"].nil?
@@ -41,6 +49,10 @@ module DockerComposeEnv
       end
     end
   rescue Errno::ENOENT => exception
-    raise exception unless exception.message == 'No such file or directory - docker-compose'
+    if exception.message == 'No such file or directory - docker-compose'
+      $stderr.puts(exception.message)
+    else
+      raise exception
+    end
   end
 end